clusplot {cluster}	R Documentation

Cluster Plot - Generic Function

Description

Draws a 2-dimensional “clusplot” on the current graphics device. This is a generic function with a default and partition method.

Usage

clusplot(x, ...)

Arguments

x 	 an R object.

... 	 additional arguments for methods. Graphical parameters (see par) may also be supplied as arguments to this function.

Side Effects

a 2-dimensional clusplot is created on the current graphics device.

See Also

(for references and examples) clusplot.default, clusplot.partition, partition.object, pam, fanny, clara.

[Package cluster version 1.12.0 Index]