diana.object {cluster}	R Documentation

Divisive Analysis (DIANA) Object

Description

The objects of class "diana" represent a divisive hierarchical clustering of a dataset.

Value

A legitimate diana object is a list with the following components:

order 	 a vector giving a permutation of the original observations to allow for plotting, in the sense that the branches of a clustering tree will not cross.

order.lab 	 a vector similar to order, but containing observation labels instead of observation numbers. This component is only available if the original observations were labelled.

height 	 a vector with the diameters of the clusters prior to splitting.

dc 	 the divisive coefficient, measuring the clustering structure of the dataset. For each observation i, denote by d(i) the diameter of the last cluster to which it belongs (before being split off as a single observation), divided by the diameter of the whole dataset. The dc is the average of all 1 - d(i). It can also be seen as the average width (or the percentage filled) of the banner plot. Because dc grows with the number of observations, this measure should not be used to compare datasets of very different sizes.

merge 	 an (n-1) by 2 matrix, where n is the number of observations. Row i of merge describes the split at step n-i of the clustering. If a number j in row r is negative, then the single observation |j| is split off at stage n-r. If j is positive, then the cluster that will be splitted at stage n-j (described by row j), is split off at stage n-r.

diss 	 an object of class "dissimilarity", representing the total dissimilarity matrix of the dataset.

data 	 a matrix containing the original or standardized measurements, depending on the stand option of the function agnes. If a dissimilarity matrix was given as input structure, then this component is not available.

GENERATION

This class of objects is returned from diana.

METHODS

The "diana" class has methods for the following generic functions: print, summary, plot.

INHERITANCE

The class "diana" inherits from "twins". Therefore, the generic function pltree can be used on a diana object, and an as.hclust method is available.

See Also

agnes, diana, plot.diana, twins.object.

Examples

## really see example(diana) !   Additionally:
data(votes.repub)
dv0 <- diana(votes.repub, stand = TRUE)
## Cut into 2 groups:
dv2 <- cutree(as.hclust(dv0), k = 2)
table(dv2)
rownames(votes.repub)[dv2 == 1]

[Package cluster version 1.9.8 Index]