Flower Characteristics

8 characteristics for 18 popular flowers.

data(flower)

Format

A data frame with 18 observations on 8 variables:

[ , "V1"]	factor	winters

[ , "V2"]	factor	shadow

[ , "V3"]	factor	tubers

[ , "V4"]	factor	color

[ , "V5"]	ordered	soil

[ , "V6"]	ordered	preference

[ , "V7"]	numeric	height

[ , "V8"]	numeric	distance

V1

winters, is binary and indicates whether the plant may be left in the garden when it freezes.

V2

shadow, is binary and shows whether the plant needs to stand in the shadow.

V3

tubers, is asymmetric binary and distinguishes between plants with tubers and plants that grow in any other way.

V4

color, is nominal and specifies the flower's color (1 = white, 2 = yellow, 3 = pink, 4 = red, 5 = blue).

V5

soil, is ordinal and indicates whether the plant grows in dry (1), normal (2), or wet (3) soil.

V6

preference, is ordinal and gives someone's preference ranking going from 1 to 18.

V7

height, is interval scaled, the plant's height in centimeters.

V8

distance, is interval scaled, the distance in centimeters that should be left between the plants.

References

Struyf, Hubert and Rousseeuw (1996), see agnes.

Examples

data(flower)
## Example 2 in ref
daisy(flower, type = list(asymm = 3))

#> Dissimilarities :
#>            1         2         3         4         5         6         7
#> 2  0.8875408                                                            
#> 3  0.5272467 0.5882353                                                  
#> 4  0.3517974 0.5504493 0.5651552                                        
#> 5  0.4115605 0.7115780 0.4258637 0.6383578                              
#> 6  0.2269199 0.7549953 0.3432306 0.4189951 0.3935574                    
#> 7  0.2876225 0.6856209 0.5595705 0.3435866 0.4797386 0.2163399          
#> 8  0.4234069 0.4641340 0.6038399 0.2960376 0.4673203 0.5714869 0.4107843
#> 9  0.5808824 0.4933240 0.5101307 0.8076797 0.3779178 0.5870682 0.6732493
#> 10 0.6094363 0.5178338 0.5346405 0.5570670 0.4357610 0.4707750 0.6703081
#> 11 0.3278595 0.7096814 0.5993873 0.6518791 0.3864788 0.4828840 0.5652369
#> 12 0.4267565 0.5857843 0.6004902 0.5132761 0.5000817 0.5248366 0.6391340
#> 13 0.5196487 0.5998133 0.6166200 0.7464461 0.3336134 0.5170868 0.6032680
#> 14 0.2926062 0.5949346 0.6096405 0.3680147 0.5203431 0.3656863 0.5049837
#> 15 0.6221814 0.4461251 0.6057890 0.5531454 0.5259570 0.5818861 0.3823529
#> 16 0.6935866 0.4085901 0.7111111 0.3417892 0.8344538 0.5837535 0.4975724
#> 17 0.7765114 0.2176471 0.6630252 0.4247141 0.7863679 0.6785247 0.5923436
#> 18 0.4610294 0.4515114 0.7162173 0.4378268 0.4755310 0.6438317 0.4692402
#>            8         9        10        11        12        13        14
#> 2                                                                       
#> 3                                                                       
#> 4                                                                       
#> 5                                                                       
#> 6                                                                       
#> 7                                                                       
#> 8                                                                       
#> 9  0.6366422                                                            
#> 10 0.6639706 0.4864146                                                  
#> 11 0.4955474 0.4308007 0.3948121                                        
#> 12 0.4216503 0.4194036 0.3812092 0.2636029                              
#> 13 0.5754085 0.2493464 0.4164332 0.3445670 0.2331699                    
#> 14 0.4558007 0.4396650 0.3609477 0.2838644 0.1591503 0.3784314          
#> 15 0.4512255 0.2908964 0.4812325 0.4806781 0.4295752 0.3637722 0.4351307
#> 16 0.6378268 0.7422502 0.3986928 0.7436683 0.6050654 0.6722689 0.4598039
#> 17 0.4707516 0.6941643 0.3506069 0.7015931 0.5629902 0.6241830 0.5427288
#> 18 0.1417892 0.5198529 0.8057598 0.5359477 0.5495507 0.5733252 0.5698121
#>           15        16        17
#> 2                               
#> 3                               
#> 4                               
#> 5                               
#> 6                               
#> 7                               
#> 8                               
#> 9                               
#> 10                              
#> 11                              
#> 12                              
#> 13                              
#> 14                              
#> 15                              
#> 16 0.4513539                    
#> 17 0.4032680 0.1909430          
#> 18 0.5096814 0.7796160 0.6125408
#> 
#> Metric :  mixed ;  Types = N, N, A, N, O, O, I, I 
#> Number of objects : 18

daisy(flower, type = list(asymm = c(1, 3), ordratio = 7))

#> Dissimilarities :
#>            1         2         3         4         5         6         7
#> 2  0.9007353                                                            
#> 3  0.6176471 0.5882353                                                  
#> 4  0.4226891 0.5455882 0.6403361                                        
#> 5  0.4806723 0.7369748 0.5264706 0.7605042                              
#> 6  0.2823529 0.7470588 0.3911765 0.4764706 0.4980392                    
#> 7  0.3310924 0.6983193 0.6676471 0.4109244 0.5745098 0.2764706          
#> 8  0.5100840 0.4544118 0.6789916 0.3327731 0.5705882 0.6563025 0.4932773
#> 9  0.5808824 0.5084034 0.5252101 0.8257353 0.3882353 0.6100840 0.6756303
#> 10 0.6323529 0.5067227 0.5235294 0.5522059 0.4722689 0.4739496 0.6941176
#> 11 0.3389706 0.7117647 0.6014706 0.6588235 0.4066176 0.4919118 0.5742647
#> 12 0.4441176 0.5816176 0.5963235 0.5139706 0.5264706 0.5220588 0.6544118
#> 13 0.5286765 0.6252101 0.6420168 0.7735294 0.3336134 0.5504202 0.6159664
#> 14 0.3044118 0.5963235 0.6110294 0.3742647 0.5411765 0.3573529 0.5147059
#> 15 0.6242647 0.4588235 0.6184874 0.5691176 0.5386555 0.6025210 0.3823529
#> 16 0.6845588 0.3831933 0.6857143 0.3147059 0.8344538 0.5504202 0.4848739
#> 17 0.7897059 0.2176471 0.6630252 0.4198529 0.8117647 0.6705882 0.6050420
#> 18 0.5268908 0.4647059 0.8336134 0.5210084 0.5537815 0.7588235 0.5386555
#>            8         9        10        11        12        13        14
#> 2                                                                       
#> 3                                                                       
#> 4                                                                       
#> 5                                                                       
#> 6                                                                       
#> 7                                                                       
#> 8                                                                       
#> 9  0.6595588                                                            
#> 10 0.6639706 0.5126050                                                  
#> 11 0.5073529 0.4419118 0.4066176                                        
#> 12 0.4272059 0.4367647 0.3867647 0.2698529                              
#> 13 0.6073529 0.2596639 0.4529412 0.3647059 0.2595588                    
#> 14 0.4669118 0.4514706 0.3720588 0.2845588 0.1647059 0.3992647          
#> 15 0.4720588 0.2932773 0.5050420 0.4897059 0.4448529 0.3764706 0.4448529
#> 16 0.6058824 0.7319328 0.3621849 0.7235294 0.5786765 0.6722689 0.4389706
#> 17 0.4610294 0.7092437 0.3394958 0.7036765 0.5588235 0.6495798 0.5441176
#> 18 0.1882353 0.5198529 0.8286765 0.5470588 0.5669118 0.5823529 0.5816176
#>           15        16        17
#> 2                               
#> 3                               
#> 4                               
#> 5                               
#> 6                               
#> 7                               
#> 8                               
#> 9                               
#> 10                              
#> 11                              
#> 12                              
#> 13                              
#> 14                              
#> 15                              
#> 16 0.4386555                    
#> 17 0.4159664 0.1655462          
#> 18 0.5117647 0.7705882 0.6257353
#> 
#> Metric :  mixed ;  Types = A, N, A, N, O, O, T, I 
#> Number of objects : 18