partition.object {cluster}	R Documentation

Partitioning Object

Description

The objects of class "partition" represent a partitioning of a dataset into clusters.

Value

a "partition" object is a list with the following (and typically more) components:

clustering
the clustering vector. An integer vector of length n, the number of observations, giving for each observation the number (‘id’) of the cluster to which it belongs.

call
the matched call generating the object.

silinfo
a list with all silhouette information, only available when the number of clusters is non-trivial, i.e., 1 < k < n and then has the following components, see silhouette

widths

an (n x 3) matrix, as returned by silhouette(), with for each observation i the cluster to which i belongs, as well as the neighbor cluster of i (the cluster, not containing i, for which the average dissimilarity between its observations and i is minimal), and the silhouette width s(i) of the observation.

clus.avg.widths

the average silhouette width per cluster.

avg.width

the average silhouette width for the dataset, i.e., simply the average of s(i) over all observations i.

This information is also needed to construct a silhouette plot of the clustering, see plot.partition.

Note that avg.width can be maximized over different clusterings (e.g. with varying number of clusters) to choose an optimal clustering.

objective
value of criterion maximized during the partitioning algorithm, may more than one entry for different stages.

diss
an object of class "dissimilarity", representing the total dissimilarity matrix of the dataset (or relevant subset, e.g. for clara).

data
a matrix containing the original or standardized data. This might be missing to save memory or when a dissimilarity matrix was given as input structure to the clustering method.

GENERATION

These objects are returned from pam, clara or fanny.

METHODS

The "partition" class has a method for the following generic functions: plot, clusplot.

INHERITANCE

The following classes inherit from class "partition" : "pam", "clara" and "fanny".

See pam.object, clara.object and fanny.object for details.

See Also

pam, clara, fanny.

[Package cluster version 2.0.7-1 Index]