plot.mona {cluster}	R Documentation

Banner of Monothetic Divisive Hierarchical Clusterings

Description

Creates the banner of a mona object.

Usage


## S3 method for class 'mona'
plot(x, main = paste("Banner of ", deparse(x$call)),
          sub = NULL, xlab = "Separation step",
	  col = c(2,0), axes = TRUE, adj = 0,
          nmax.lab = 35, max.strlen = 5, ...)

Arguments

x
an object of class "mona", typically created by mona(.).

main,sub
main and sub titles for the plot, with convenient defaults. See documentation in plot.default.

xlab
x axis label, see title.

col,adj
graphical parameters passed to bannerplot().

axes
logical, indicating if (labeled) axes should be drawn.

nmax.lab
integer indicating the number of labels which is considered too large for labeling.

max.strlen
positive integer giving the length to which strings are truncated in labeling.

...
further graphical arguments are passed to bannerplot() and text.

Details

Plots the separation step at which clusters are splitted. The observations are given in the order found by the mona algorithm, the numbers in the step vector are represented as bars between the observations.

When a long bar is drawn between two observations, those observations have the same value for each variable. See chapter 7 of Kaufman and Rousseeuw (1990).

Side Effects

A banner is plotted on the current graphics device.

Note

In the banner plot, observation labels are only printed when the number of observations is limited less than nmax.lab (35, by default), for readability. Moreover, observation labels are truncated to maximally max.strlen (5) characters.

References

see those in plot.agnes.

See Also

mona, mona.object, par.

[Package cluster version 2.0.7-1 Index]