pltree {cluster}	R Documentation

Clustering Trees - Generic Function

Description

Generic function drawing a clustering tree (``dendrogram'') on the current graphics device. There is Functions with names beginning in pltree. will be methods for this function.

Usage

pltree(x, ...)

Arguments

x 	 an R object (for which a pltree method is defined).

... 	 methods may have additional arguments. Graphical parameters (see par) may also be supplied as arguments to this function.

Side Effects

a clustering tree is created on the current graphics device.

See Also

twins.object, agnes, diana, par.

[Package cluster version 1.9.8 Index]