pltree.twins {cluster}	R Documentation

Clustering Tree of a Hierarchical Clustering

Description

Draws a clustering tree (dendrogramm) of a twins object, i.e., hierarchical clustering, typically resulting from agnes or diana.

Usage

## S3 method for class 'twins':
pltree(x, main = paste("Dendrogram of ", deparse(x$call)),
             labels = NULL, ylab = "Height", ...)

Arguments

x 	 an object of class "twins", typically created by either agnes() or diana().

main 	 main title with a sensible default.

labels 	 labels to use; the default is constructed from x.

ylab 	 label for y-axis.

... 	 Graphical parameters (see par) may also be supplied as arguments to this function.

Details

Creates a plot of a clustering tree given a twins object. The leaves of the tree are the original observations. In case of an agglomerative clustering, two branches come together at the distance between the two clusters being merged. For a divisive clustering, a branch splits up at the diameter of the cluster being splitted.

Value

a NULL value is returned.

See Also

agnes, agnes.object, diana, diana.object, hclust, par, plot.agnes, plot.diana.

Examples

[Package cluster version 1.9.8 Index]