votes.repub {cluster}	R Documentation

Votes for Republican Candidate in Presidential Elections

Description

A data frame with the percents of votes given to the republican candidate in presidential elections from 1856 to 1976. Rows represent the 50 states, and columns the 31 elections.

Usage

data(votes.repub)

Source

S. Peterson (1973): A Statistical History of the American Presidential Elections. New York: Frederick Ungar Publishing Co.

Data from 1964 to 1976 is from R. M. Scammon, American Votes 12, Congressional Quarterly.

[Package cluster version 1.12.0 Index]