xclara {cluster}	R Documentation

Bivariate Data Set with 3 Clusters

Description

An artificial data set consisting of 3000 points in 3 well-separated clusters of size 1000 each.

Usage

data(xclara)

Format

A data frame with 3000 observations on 2 numeric variables giving the x and y coordinates of the points, respectively.

Source

Sample data set accompanying the reference below, obtained from http://www.stat.ucla.edu/journals/jss/v01/i04/data/.

References

Anja Struyf, Mia Hubert & Peter J. Rousseeuw (1996): Clustering in an Object-Oriented Environment. Journal of Statistical Software, 1. http://www.stat.ucla.edu/journals/jss/

[Package cluster version 1.12.0 Index]