clara.object {cluster}	R Documentation

Clustering Large Applications (CLARA) Object

Description

The objects of class "clara" represent a partitioning of a large dataset into clusters and are typically returned from clara.

Value

A legitimate clara object is a list with the following components:

sample 	 labels or case numbers of the observations in the best sample, that is, the sample used by the clara algorithm for the final partition.

medoids 	 the medoids or representative objects of the clusters. It is a matrix with in each row the coordinates of one medoid. Possibly NULL, namely when the object resulted from clara(*, medoids.x=FALSE). Use the following i.med in that case.

i.med 	 the indices of the medoids above: medoids <- x[i.med,] where x is the original data matrix in clara(x,*).

clustering 	 the clustering vector, see partition.object.

objective 	 the objective function for the final clustering of the entire dataset.

clusinfo 	 matrix, each row gives numerical information for one cluster. These are the cardinality of the cluster (number of observations), the maximal and average dissimilarity between the observations in the cluster and the cluster's medoid. The last column is the maximal dissimilarity between the observations in the cluster and the cluster's medoid, divided by the minimal dissimilarity between the cluster's medoid and the medoid of any other cluster. If this ratio is small, the cluster is well-separated from the other clusters.

diss 	 dissimilarity (maybe NULL), see partition.object.

silinfo 	 list with silhouette width information for the best sample, see partition.object.

call 	 generating call, see partition.object.

data 	 matrix, possibibly standardized, or NULL, see partition.object.

Methods, Inheritance

The "clara" class has methods for the following generic functions: print, summary.

The class "clara" inherits from "partition". Therefore, the generic functions plot and clusplot can be used on a clara object.

See Also

clara, dissimilarity.object, partition.object, plot.partition.

[Package cluster version 1.12.0 Index]