clusplot.partition {cluster}	R Documentation

Bivariate Clusplot of a Partitioning Object

Description

Clusplot (Clustering Plot) method for an object of class partition.

Usage

## S3 method for class 'partition':
clusplot(x, main = NULL, dist = NULL, ...)

Arguments

x 	 an object of class "partition", e.g. created by the functions pam, clara, or fanny.

main 	 title for the plot; when NULL (by default), a title is constructed, using x$call.

dist 	 when x does not have a diss nor a data component, e.g., for pam(dist(*), keep.diss=FALSE), dist must specify the dissimilarity for the clusplot.

... 	 all optional arguments available for the clusplot.default function (except for the diss one) may also be supplied to this function. Graphical parameters (see par) may also be supplied as arguments to this function.

Details

This clusplot.partition() method relies on clusplot.default.

If the clustering algorithms pam, fanny and clara are applied to a data matrix of observations-by-variables then a clusplot of the resulting clustering can always be drawn. When the data matrix contains missing values and the clustering is performed with pam or fanny, the dissimilarity matrix will be given as input to clusplot. When the clustering algorithm clara was applied to a data matrix with NAs then clusplot will replace the missing values as described in clusplot.default, because a dissimilarity matrix is not available.

Value

An invisible list with components

Distances 	 When option lines is 1 or 2 we optain a k by k matrix (k is the number of clusters). The element at row j and column s is the distance between ellipse j and ellipse s. If lines=0, then the value of this component is NA.

Shading 	 A vector of length k (where k is the number of clusters), containing the amount of shading per cluster. Let y be a vector where element i is the ratio between the number of objects in cluster i and the area of ellipse i. When the cluster i is a line segment, y[i] and the density of the cluster are set to NA. Let z be the sum of all the elements of y without the NAs. Then we put shading = y/z *37 + 3.

See Also

clusplot.default for references; partition.object, pam, pam.object, clara, clara.object, fanny, fanny.object, par.

Examples

## generate 25 objects, divided into 2 clusters.
x <- rbind(cbind(rnorm(10,0,0.5), rnorm(10,0,0.5)),
           cbind(rnorm(15,5,0.5), rnorm(15,5,0.5)))
clusplot(pam(x, 2))
## add noise, and try again :
x4 <- cbind(x, rnorm(25), rnorm(25))
clusplot(pam(x4, 2))

[Package cluster version 1.12.0 Index]