Dissimilarity Matrix Object

Objects of class "dissimilarity" representing the dissimilarity matrix of a dataset.

GENERATION

daisy returns this class of objects. Also the functions pam, clara, fanny, agnes, and diana return a dissimilarity object, as one component of their return objects.

METHODS

The "dissimilarity" class has methods for the following generic functions: print, summary.

Value

The dissimilarity matrix is symmetric, and hence its lower triangle (column wise) is represented as a vector to save storage space. If the object, is called do, and n the number of observations, i.e., n <- attr(do, "Size"), then for \(i < j <= n\), the dissimilarity between (row) i and j is do[n*(i-1) - i*(i-1)/2 + j-i]. The length of the vector is \(n*(n-1)/2\), i.e., of order \(n^2\).

"dissimilarity" objects also inherit from class dist and can use dist methods, in particular, as.matrix, such that \(d_{ij}\) from above is just as.matrix(do)[i,j].

The object has the following attributes:

Size

the number of observations in the dataset.

Metric

the metric used for calculating the dissimilarities. Possible values are "euclidean", "manhattan", "mixed" (if variables of different types were present in the dataset), and "unspecified".

Labels

optionally, contains the labels, if any, of the observations of the dataset.

NA.message

optionally, if a dissimilarity could not be computed, because of too many missing values for some observations of the dataset.

Types

when a mixed metric was used, the types for each variable as one-letter codes (as in the book, e.g. p.54):

A

Asymmetric binary

S

Symmetric binary

N

Nominal (factor)

O

Ordinal (ordered factor)

I

Interval scaled (numeric)

T

raTio to be log transformed (positive numeric)

.

See also