Fuzzy Analysis (FANNY) Object

The objects of class "fanny" represent a fuzzy clustering of a dataset.

GENERATION

These objects are returned from fanny.

METHODS

The "fanny" class has methods for the following generic functions: print, summary.

INHERITANCE

The class "fanny" inherits from "partition". Therefore, the generic functions plot and clusplot can be used on a fanny object.

Value

A legitimate fanny object is a list with the following components:

membership

matrix containing the memberships for each pair consisting of an observation and a cluster.

memb.exp

the membership exponent used in the fitting criterion.

coeff

Dunn's partition coefficient \(F(k)\) of the clustering, where \(k\) is the number of clusters. \(F(k)\) is the sum of all squared membership coefficients, divided by the number of observations. Its value is between \(1/k\) and 1. The normalized form of the coefficient is also given. It is defined as \((F(k) - 1/k) / (1 - 1/k)\), and ranges between 0 and 1. A low value of Dunn's coefficient indicates a very fuzzy clustering, whereas a value close to 1 indicates a near-crisp clustering.

clustering

the clustering vector of the nearest crisp clustering, see partition.object.

k.crisp

integer (\(\le k\)) giving the number of crisp clusters; can be less than \(k\), where it's recommended to decrease memb.exp.

objective

named vector containing the minimal value of the objective function reached by the FANNY algorithm and the relative convergence tolerance tol used.

convergence

named vector with iterations, the number of iterations needed and converged indicating if the algorithm converged (in maxit iterations within convergence tolerance tol).

diss

an object of class "dissimilarity", see partition.object.

call

generating call, see partition.object.

silinfo

list with silhouette information of the nearest crisp clustering, see partition.object.

data

matrix, possibibly standardized, or NULL, see partition.object.

See also