mona.object {cluster}	R Documentation

Monothetic Analysis (MONA) Object

Description

The objects of class "mona" represent the divisive hierarchical clustering of a dataset with only binary variables (measurements). This class of objects is returned from mona.

Value

A legitimate mona object is a list with the following components:

data 	 matrix with the same dimensions as the original data matrix, but with factors coded as 0 and 1, and all missing values replaced.

order 	 a vector giving a permutation of the original observations to allow for plotting, in the sense that the branches of a clustering tree will not cross.

order.lab 	 a vector similar to order, but containing observation labels instead of observation numbers. This component is only available if the original observations were labelled.

variable 	 vector of length n-1 where n is the number of observations, specifying the variables used to separate the observations of order.

step 	 vector of length n-1 where n is the number of observations, specifying the separation steps at which the observations of order are separated.

METHODS

The "mona" class has methods for the following generic functions: print, summary, plot.

See Also

mona for examples etc, plot.mona.

[Package cluster version 1.12.0 Index]