print.agnes {cluster}	R Documentation

Print Method for AGNES Objects

Description

Prints the call, agglomerative coefficient, ordering of objects and distances between merging clusters (`Height') of an agnes object.

This is a method for the generic print() function for objects inheriting from class agnes, see agnes.object.

Usage

## S3 method for class 'agnes':
print(x, ...)

Arguments

x 	 an agnes object.

... 	 potential further arguments (required by generic).

See Also

summary.agnes producing more output; agnes, agnes.object, print, print.default.

[Package cluster version 1.12.0 Index]