sizeDiss {cluster}	R Documentation

Sample Size of Dissimilarity Like Object

Description

Returns the number of observations (sample size) corresponding to a dissimilarity like object, or equivalently, the number of rows or columns of a matrix when only the lower or upper triangular part (without diagonal) is given.

It is nothing else but the inverse function of f(n) = n(n-1)/2.

Usage

sizeDiss(d)

Arguments

d 	 any R object with length (typically) n(n-1)/2.

Value

a number; n if length(d) == n(n-1)/2, NA otherwise.

See Also

dissimilarity.object and also as.dist for class dissimilarity and dist objects which have a Size attribute.

Examples

sizeDiss(1:10)# 5, since 10 == 5 * (5 - 1) / 2
sizeDiss(1:9) # NA

n <- 1:100
stopifnot(n == sapply( n*(n-1)/2, function(n) sizeDiss(logical(n))))

[Package cluster version 1.12.0 Index]