volume.ellipsoid {cluster}	R Documentation

Compute the Volume of Planar Object

Description

Compute the volume of a planar object. This is a generic function and a method for ellipsoid objects.

Usage

## S3 method for class 'ellipsoid':
volume(object)

Arguments

object 	 an R object the volume of which is wanted; for the ellipsoid method, an object of that class (see ellipsoidhull or the example below).

Value

a number, the volume of the given object.

See Also

ellipsoidhull for spanning ellipsoid computation.

Examples

## example(ellipsoidhull) # which defines `ellipsoid' object <namefoo>

myEl <- structure(list(cov = rbind(c(3,1),1:2), loc = c(0,0), d2 = 10),
                   class = "ellipsoid")
volume(myEl)# i.e. "area" here (d = 2)
myEl # also mentions the "volume"

[Package cluster version 1.12.0 Index]