volume.ellipsoid {cluster}	R Documentation

Compute the Volume (of an Ellipsoid)

Description

Compute the volume of geometric R object. This is a generic function and has a method for ellipsoid objects (typically resulting from ellipsoidhull().

Usage

volume(object, ...)
## S3 method for class 'ellipsoid'
volume(object, log = FALSE, ...)

Arguments

object
an R object the volume of which is wanted; for the ellipsoid method, an object of that class (see ellipsoidhull or the example below).

log
logical indicating if the volume should be returned in log scale. Maybe needed in largish dimensions.

...
potential further arguments of methods, e.g. log.

Value

a number, the volume V (or \log(V) if log = TRUE) of the given object.

Author(s)

Martin Maechler (2002, extracting from former clusplot code); Keefe Murphy (2019) provided code for dimensions d > 2.

See Also

ellipsoidhull for spanning ellipsoid computation.

Examples

## example(ellipsoidhull) # which defines 'ellipsoid' object <namefoo>

myEl <- structure(list(cov = rbind(c(3,1),1:2), loc = c(0,0), d2 = 10),
                   class = "ellipsoid")
volume(myEl)# i.e. "area" here (d = 2)
myEl # also mentions the "volume"

set.seed(1)
d5 <- matrix(rt(500, df=3), 100,5)
e5 <- ellipsoidhull(d5)

[Package cluster version 2.1.0 Index]